updates to exercise 12 answer
Got the basic premise working but didn’t flesh it out as much as he has
in the exercise solutions. The biggest thing was that I needed to
change the way the gradients were computed, specifically to use /m
instead of *(1/m) for element wise matrix division by a constant. Not
sure why the second method didn’t work, but it produced matrices full
of zeros.